Gradient-Backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
brotli==1.2.0
cachetools==5.5.2
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comtypes==1.4.16
cryptography==45.0.6
ddgs==9.10.0
distro==1.9.0
dnspython==2.6.1
duckdb==1.4.1
duckduckgo_search==8.1.1
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fake-useragent==2.2.0
fastapi==0.121.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
lxml==6.0.2
oauthlib==3.3.1
openai==2.16.0
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
primp==0.15.0
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
pypiwin32==223
pytest==8.3.4
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.26
pyttsx3==2.99
pywin32==311
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
starlette==0.44.0
tqdm==4.67.2
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.33.0
